+++ pkg/model/database/database.go
@@ -36,6 +36,13 @@ type TimeMetadata struct {
 	UpdatedAt *time.Time `gorm:"nullable"`
 }
 
+// SafeSQLName represents a value that is safe to use as an SQL table or column
+// name without the need of escaping.
+//
+// It is merely semantical and has no validation attached. Values of this type
+// should never be constructed from user input.
+type SafeSQLName string
+
 // ProviderFields holds the Go struct field names for each field.
 // Useful in GORM .Where() statements to only select certain fields or in GORM
 // Preload statements to select the correct field to preload.
@@ -85,6 +92,7 @@ type Token struct {
 // Preload statements to select the correct field to preload.
 var ProjectFields = struct {
 	ProjectID       string
+	RemoteProjectID string
 	Name            string
 	GroupName       string
 	Description     string
@@ -115,19 +123,21 @@ var ProjectFields = struct {
 // Useful in GORM .Order() statements to order the results based on a specific
 // column, which does not support the regular Go field names.
 var ProjectColumns = struct {
-	ProjectID   string
-	Name        string
-	GroupName   string
-	Description string
-	TokenID     string
-	GitURL      string
+	ProjectID       SafeSQLName
+	RemoteProjectID SafeSQLName
+	Name            SafeSQLName
+	GroupName       SafeSQLName
+	Description     SafeSQLName
+	TokenID         SafeSQLName
+	GitURL          SafeSQLName
 }{
-	ProjectID:   "project_id",
-	Name:        "name",
-	GroupName:   "group_name",
-	Description: "description",
-	TokenID:     "token_id",
-	GitURL:      "git_url",
+	ProjectID:       "project_id",
+	RemoteProjectID: "remote_project_id",
+	Name:            "name",
+	GroupName:       "group_name",
+	Description:     "description",
+	TokenID:         "token_id",
+	GitURL:          "git_url",
 }
 
 // Project holds data about an imported project. A lot of the data is expected
@@ -136,6 +146,7 @@ var ProjectColumns = struct {
 type Project struct {
 	TimeMetadata
 	ProjectID       uint      `gorm:"primaryKey"`
+	RemoteProjectID string    `gorm:"not null;default:''"`
 	Name            string    `gorm:"size:500;not null"`
 	GroupName       string    `gorm:"size:500;not null;default:''"`
 	Description     string    `gorm:"size:500;not null;default:''"`
@@ -168,8 +179,8 @@ var BranchFields = struct {
 // Useful in GORM .Order() statements to order the results based on a specific
 // column, which does not support the regular Go field names.
 var BranchColumns = struct {
-	BranchID string
-	Name     string
+	BranchID SafeSQLName
+	Name     SafeSQLName
 }{
 	BranchID: "branch_id",
 	Name:     "name",
@@ -203,14 +214,14 @@ var BuildFields = struct {
 // Useful in GORM .Order() statements to order the results based on a specific
 // column, which does not support the regular Go field names.
 var BuildColumns = struct {
-	BuildID     string
-	StatusID    string
-	ScheduledOn string
-	StartedOn   string
-	CompletedOn string
-	Environment string
-	Stage       string
-	IsInvalid   string
+	BuildID     SafeSQLName
+	StatusID    SafeSQLName
+	ScheduledOn SafeSQLName
+	StartedOn   SafeSQLName
+	CompletedOn SafeSQLName
+	Environment SafeSQLName
+	Stage       SafeSQLName
+	IsInvalid   SafeSQLName
 }{
 	BuildID:     "build_id",
 	StatusID:    "status_id",
@@ -317,8 +328,8 @@ type Param struct {
 // Useful in GORM .Order() statements to order the results based on a specific
 // column, which does not support the regular Go field names.
 var ArtifactColumns = struct {
-	ArtifactID string
-	FileName   string
+	ArtifactID SafeSQLName
+	FileName   SafeSQLName
 }{
 	ArtifactID: "artifact_id",
 	FileName:   "file_name",