+++ pkg/model/response/response.go
@@ -153,23 +153,26 @@ type Ping struct {
 // Useful in ordering statements to map the correct field to the correct
 // database column.
 var ProjectJSONFields = struct {
-	ProjectID   string
-	Name        string
-	GroupName   string
-	Description string
-	GitURL      string
+	ProjectID       string
+	RemoteProjectID string
+	Name            string
+	GroupName       string
+	Description     string
+	GitURL          string
 }{
-	ProjectID:   "projectId",
-	Name:        "name",
-	GroupName:   "groupName",
-	Description: "description",
-	GitURL:      "gitUrl",
+	ProjectID:       "projectId",
+	RemoteProjectID: "remoteProjectId",
+	Name:            "name",
+	GroupName:       "groupName",
+	Description:     "description",
+	GitURL:          "gitUrl",
 }
 
 // Project holds details about a project.
 type Project struct {
 	TimeMetadata
 	ProjectID             uint        `json:"projectId" minimum:"0"`
+	RemoteProjectID       string      `json:"remoteProjectId"`
 	Name                  string      `json:"name"`
 	GroupName             string      `json:"groupName"`
 	Description           string      `json:"description"`